+++ services/web/src/components/ui/rate/index.ts
@@ -1,3 +1,4 @@
 import Rate from "antd/lib/rate";
+export * from "antd/lib/rate";
 
 export default Rate;
