+++ services/web/src/features/bookshelf/index.ts
@@ -1,3 +1,4 @@
 export * from "./routes/AddBook";
 export * from "./routes/BookDetail";
 export * from "./routes/BookshelfList";
+export * from "./routes/Dashboard";
